refactor base interactor
