+++ novawallet/Modules/Staking/Mythos/MythosStakingSetup/MythosStakingSetupInteractor.swift
@@ -2,33 +2,24 @@ import UIKit
 import SubstrateSdk
 import Operation_iOS
 
-final class MythosStakingSetupInteractor: RuntimeConstantFetching {
-    weak var presenter: MythosStakingSetupInteractorOutputProtocol?
+final class MythosStakingSetupInteractor: MythosStakingBaseInteractor {
+    var presenter: MythosStakingSetupInteractorOutputProtocol? {
+        get {
+            basePresenter as? MythosStakingSetupInteractorOutputProtocol
+        }
 
-    let chainAsset: ChainAsset
-    let selectedAccount: ChainAccountResponse
-    let walletLocalSubscriptionFactory: WalletLocalSubscriptionFactoryProtocol
-    let priceLocalSubscriptionFactory: PriceProviderFactoryProtocol
+        set {
+            basePresenter = newValue
+        }
+    }
+
     let preferredCollatorFactory: PreferredStakingCollatorFactoryProtocol?
-    let generalLocalSubscriptionFactory: GeneralStorageSubscriptionFactoryProtocol
     let rewardService: CollatorStakingRewardCalculatorServiceProtocol
-    let frozenBalanceStore: MythosStakingFrozenBalanceStore
-    let stakingDetailsService: MythosStakingDetailsSyncServiceProtocol
-    let extrinsicService: ExtrinsicServiceProtocol
-    let feeProxy: ExtrinsicFeeProxyProtocol
     let repositoryFactory: SubstrateRepositoryFactoryProtocol
     let connection: JSONRPCEngine
-    let runtimeProvider: RuntimeCodingServiceProtocol
-    let stakingLocalSubscriptionFactory: MythosStakingLocalSubscriptionFactoryProtocol
     let identityProxyFactory: IdentityProxyFactoryProtocol
-    let operationQueue: OperationQueue
-    let logger: LoggerProtocol
 
-    private var balanceProvider: StreamableProvider<AssetBalance>?
-    private var minStakeProvider: AnyDataProvider<DecodedBigUInt>?
-    private var priceProvider: StreamableProvider<PriceData>?
     private var collatorSubscription: CallbackStorageSubscription<MythosStakingPallet.CandidateInfo>?
-    private var blockNumberProvider: AnyDataProvider<DecodedBlockNumber>?
     private var delegatorIdentityCancellable = CancellableCallStore()
 
     private lazy var localKeyFactory = LocalStorageKeyFactory()
@@ -53,114 +44,50 @@ final class MythosStakingSetupInteractor: RuntimeConstantFetching {
         operationQueue: OperationQueue,
         logger: LoggerProtocol
     ) {
-        self.chainAsset = chainAsset
-        self.selectedAccount = selectedAccount
-        self.stakingDetailsService = stakingDetailsService
-        self.rewardService = rewardService
-        self.stakingLocalSubscriptionFactory = stakingLocalSubscriptionFactory
-        self.walletLocalSubscriptionFactory = walletLocalSubscriptionFactory
-        self.priceLocalSubscriptionFactory = priceLocalSubscriptionFactory
-        self.generalLocalSubscriptionFactory = generalLocalSubscriptionFactory
         self.preferredCollatorFactory = preferredCollatorFactory
-        self.extrinsicService = extrinsicService
-        self.feeProxy = feeProxy
-        self.connection = connection
-        self.runtimeProvider = runtimeProvider
+        self.rewardService = rewardService
         self.repositoryFactory = repositoryFactory
+        self.connection = connection
         self.identityProxyFactory = identityProxyFactory
-        self.operationQueue = operationQueue
-        self.logger = logger
 
-        frozenBalanceStore = MythosStakingFrozenBalanceStore(
-            accountId: selectedAccount.accountId,
-            chainAssetId: chainAsset.chainAssetId,
+        super.init(
+            chainAsset: chainAsset,
+            selectedAccount: selectedAccount,
+            stakingDetailsService: stakingDetailsService,
+            stakingLocalSubscriptionFactory: stakingLocalSubscriptionFactory,
             walletLocalSubscriptionFactory: walletLocalSubscriptionFactory,
+            priceLocalSubscriptionFactory: priceLocalSubscriptionFactory,
+            generalLocalSubscriptionFactory: generalLocalSubscriptionFactory,
+            extrinsicService: extrinsicService,
+            feeProxy: feeProxy,
+            runtimeProvider: runtimeProvider,
+            currencyManager: currencyManager,
+            operationQueue: operationQueue,
             logger: logger
         )
-
-        self.currencyManager = currencyManager
     }
 
     deinit {
         delegatorIdentityCancellable.cancel()
     }
 
-    func getExtrinsicBuilderClosure(from model: MythosStakeModel) -> ExtrinsicBuilderClosure {
-        { builder in
-            var resultBuilder = builder
+    override func onSetup() {
+        super.onSetup()
 
-            if model.amount.toLock > 0 {
-                let lockCall = MythosStakingPallet.LockCall(amount: model.amount.toLock)
-                resultBuilder = try resultBuilder.adding(call: lockCall.runtimeCall())
-            }
+        providePreferredCollator()
+        provideRewardCalculator()
+    }
 
-            let stakeCall = MythosStakingPallet.StakeCall(
-                targets: [
-                    MythosStakingPallet.StakeTarget(
-                        candidate: model.collator,
-                        stake: model.amount.toStake
-                    )
-                ]
-            )
+    override func onStakingDetails(_ stakingDetails: MythosStakingDetails?) {
+        super.onStakingDetails(stakingDetails)
 
-            return try resultBuilder.adding(call: stakeCall.runtimeCall())
+        if let stakingDetails, !stakingDetails.stakeDistribution.isEmpty {
+            provideIdentities(for: Array(stakingDetails.stakeDistribution.keys))
         }
     }
 }
 
 private extension MythosStakingSetupInteractor {
-    func makeAssetBalanceSubscription() {
-        balanceProvider = subscribeToAssetBalanceProvider(
-            for: selectedAccount.accountId,
-            chainId: chainAsset.chain.chainId,
-            assetId: chainAsset.asset.assetId
-        )
-    }
-
-    func makeFrozenBalanceSubscription() {
-        frozenBalanceStore.setup()
-
-        frozenBalanceStore.add(
-            observer: self,
-            sendStateOnSubscription: true,
-            queue: .main
-        ) { [weak self] _, newState in
-            if let frozenBalance = newState {
-                self?.presenter?.didReceiveFrozenBalance(frozenBalance)
-            }
-        }
-    }
-
-    func makeMinStakeSubscription() {
-        minStakeProvider = subscribeToMinStake(for: chainAsset.chain.chainId)
-    }
-
-    func makePriceSubscription() {
-        if let priceId = chainAsset.asset.priceId {
-            priceProvider?.removeObserver(self)
-
-            priceProvider = subscribeToPrice(for: priceId, currency: selectedCurrency)
-        }
-    }
-
-    func makeStakingDetailsSubscription() {
-        stakingDetailsService.add(
-            observer: self,
-            sendStateOnSubscription: true,
-            queue: .main
-        ) { [weak self] _, newState in
-            if let newState, !newState.stakeDistribution.isEmpty {
-                self?.provideIdentities(for: Array(newState.stakeDistribution.keys))
-            }
-
-            self?.presenter?.didReceiveDetails(newState)
-        }
-    }
-
-    func makeBlockNumberSubscription() {
-        blockNumberProvider = subscribeToBlockNumber(for: chainAsset.chain.chainId)
-    }
-
     func provideRewardCalculator() {
         let operation = rewardService.fetchCalculatorOperation()
 
@@ -259,134 +186,10 @@ private extension MythosStakingSetupInteractor {
             }
         }
     }
-
-    func provideMaxCandidatesPerStaker() {
-        fetchConstant(
-            for: MythosStakingPallet.maxStakedCandidatesPath,
-            runtimeCodingService: runtimeProvider,
-            operationManager: OperationManager(operationQueue: operationQueue)
-        ) { [weak self] (result: Result<UInt32, Error>) in
-            switch result {
-            case let .success(maxCandidatesPerStaker):
-                self?.presenter?.didReceiveMaxCollatorsPerStaker(maxCandidatesPerStaker)
-            case let .failure(error):
-                self?.logger.error("Unexpected error: \(error)")
-            }
-        }
-    }
 }
 
 extension MythosStakingSetupInteractor: MythosStakingSetupInteractorInputProtocol {
-    func setup() {
-        feeProxy.delegate = self
-
-        makeAssetBalanceSubscription()
-        makePriceSubscription()
-        makeFrozenBalanceSubscription()
-        makeBlockNumberSubscription()
-
-        makeStakingDetailsSubscription()
-        makeMinStakeSubscription()
-
-        providePreferredCollator()
-        provideRewardCalculator()
-
-        provideMaxCandidatesPerStaker()
-    }
-
     func applyCollator(with accountId: AccountId) {
         subscribeRemoteCollator(for: accountId)
-    }
-
-    func estimateFee(with model: MythosStakeModel) {
-        feeProxy.estimateFee(
-            using: extrinsicService,
-            reuseIdentifier: model.reuseTxId,
-            payingIn: nil,
-            setupBy: getExtrinsicBuilderClosure(from: model)
-        )
-    }
-}
-
-extension MythosStakingSetupInteractor: ExtrinsicFeeProxyDelegate {
-    func didReceiveFee(
-        result: Result<ExtrinsicFeeProtocol, Error>,
-        for _: TransactionFeeId
-    ) {
-        switch result {
-        case let .success(model):
-            presenter?.didReceiveFee(model)
-        case let .failure(error):
-            presenter?.didReceiveError(.feeFailed(error))
-        }
-    }
-}
-
-extension MythosStakingSetupInteractor: MythosStakingLocalStorageSubscriber, MythosStakingLocalStorageHandler {
-    func handleMinStake(
-        result: Result<Balance?, Error>,
-        chainId _: ChainModel.Id
-    ) {
-        switch result {
-        case let .success(minStake):
-            if let minStake {
-                presenter?.didReceiveMinStakeAmount(minStake)
-            }
-        case let .failure(error):
-            logger.error("Min stake subscription failed: \(error)")
-        }
-    }
-}
-
-extension MythosStakingSetupInteractor: WalletLocalStorageSubscriber, WalletLocalSubscriptionHandler {
-    func handleAssetBalance(
-        result: Result<AssetBalance?, Error>,
-        accountId _: AccountId,
-        chainId _: ChainModel.Id,
-        assetId _: AssetModel.Id
-    ) {
-        switch result {
-        case let .success(balance):
-            presenter?.didReceiveAssetBalance(balance)
-        case let .failure(error):
-            logger.error("Balance subscription failed: \(error)")
-        }
-    }
-}
-
-extension MythosStakingSetupInteractor: PriceLocalStorageSubscriber, PriceLocalSubscriptionHandler {
-    func handlePrice(result: Result<PriceData?, Error>, priceId _: AssetModel.PriceId) {
-        switch result {
-        case let .success(priceData):
-            presenter?.didReceivePrice(priceData)
-        case let .failure(error):
-            logger.error("Price subscription failed: \(error)")
-        }
-    }
-}
-
-extension MythosStakingSetupInteractor: GeneralLocalStorageSubscriber, GeneralLocalStorageHandler {
-    func handleBlockNumber(
-        result: Result<BlockNumber?, Error>,
-        chainId _: ChainModel.Id
-    ) {
-        switch result {
-        case let .success(blockNumber):
-            if let blockNumber {
-                presenter?.didReceiveBlockNumber(blockNumber)
-            }
-        case let .failure(error):
-            logger.error("Unexpected error: \(error)")
-        }
-    }
-}
-
-extension MythosStakingSetupInteractor: SelectedCurrencyDepending {
-    func applyCurrency() {
-        guard presenter != nil else {
-            return
-        }
-
-        makePriceSubscription()
     }
 }
