+++ novawallet/Modules/Staking/Mythos/MythosStakingSetup/MythosStakingSetupProtocols.swift
@@ -1,25 +1,14 @@
 import BigInt
 
-protocol MythosStakingSetupInteractorInputProtocol: AnyObject {
-    func setup()
+protocol MythosStakingSetupInteractorInputProtocol: MythosStakingBaseInteractorInputProtocol {
     func applyCollator(with accountId: AccountId)
-    func estimateFee(with model: MythosStakeModel)
 }
 
-protocol MythosStakingSetupInteractorOutputProtocol: AnyObject {
-    func didReceiveAssetBalance(_ balance: AssetBalance?)
-    func didReceivePrice(_ priceData: PriceData?)
-    func didReceiveFee(_ fee: ExtrinsicFeeProtocol)
+protocol MythosStakingSetupInteractorOutputProtocol: MythosStakingBaseInteractorOutputProtocol {
     func didReceiveRewardCalculator(_ calculator: CollatorStakingRewardCalculatorEngineProtocol)
-    func didReceiveMinStakeAmount(_ amount: BigUInt)
-    func didReceiveMaxCollatorsPerStaker(_ maxCollatorsPerStaker: UInt32)
-    func didReceiveDetails(_ details: MythosStakingDetails?)
     func didReceiveCandidateInfo(_ info: MythosStakingPallet.CandidateInfo?)
     func didReceiveDelegationIdentities(_ identities: [AccountId: AccountIdentity]?)
-    func didReceiveBlockNumber(_ blockNumber: BlockNumber)
     func didReceivePreferredCollator(_ collator: DisplayAddress?)
-    func didReceiveFrozenBalance(_ frozenBalance: MythosStakingFrozenBalance)
-    func didReceiveError(_ error: MythosStakingSetupError)
 }
 
 protocol MythosStakingSetupWireframeProtocol: AlertPresentable, ErrorPresentable, FeeRetryable,
@@ -34,8 +23,4 @@ protocol MythosStakingSetupWireframeProtocol: AlertPresentable, ErrorPresentable
         from view: CollatorStakingSetupViewProtocol?,
         delegate: CollatorStakingSelectDelegate
     )
-}
-
-enum MythosStakingSetupError: Error {
-    case feeFailed(Error)
 }
